Separate type predicates from list.ts to type-predicates.ts
